requirements.txt
fastapi
uvicorn
pydantic
tortoise-orm
passlib[bcrypt]